part 01

+++ ft_putstr.c
@@ -6,19 +6,16 @@
 /*   By: dasalaza <dasalaza@student.42barcelona.c>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 18:20:36 by dasalaza          #+#    #+#             */
-/*   Updated: 2023/09/20 19:46:46 by dasalaza         ###   ########.fr       */
+/*   Updated: 2023/09/22 23:25:15 by dasalaza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//
-// Created by daruuu on 9/19/23.
-//
 #include "libftprintf.h"
 
 int ft_putstr(va_list args)
 {
     char    *str;
-    size_t  len;
+    int     len;
 
     str = va_arg(args, char *);
     len = ft_strlen(str);
